requirements.txt
google-cloud-storage
google-cloud-bigquery
google-cloud-aiplatform
pandas
matplotlib
seaborn
streamlit